`PwParser`: improve the parsing of `relax` and `vc-relax` modes
The output of relax calculations by `pw.x` can be very ambiguous in
terms of whether convergence was reached. There are various pathways
possible where the convergence thresholds were not met, but no error
messages were written or they were not properly sought for by the
parser. As a result, many failed relax calculations would unjustly be
marked as successful.