`BaseRestartWorkChain`: allow handlers to override exit status
Sometimes a calculation returns a non-zero exit code indicating it is
failed, but it can actually still be considered successful by the
workchain. In this case, the handler wants to override the exit code of
the calculation and return `ExitCode(0)` to indicate that the workchain
is successful despite the last calculation being failed. To facilitate
this the logic in the `BaseRestartWorkChain` is updated to detect zero
being returned by a handler.

In addition, to make this work, the default value of the `exit_code` for
an `ErrorHandlerReport` tuple needs to be `None` to be able to
distinguish when `ExitCode(0)` is purposefully returned.