`PwBaseWorkChain`: add error handlers for ionic calculations
Now that the `PwParser` will report trustworthy exit codes for the
successfulness of ionic calculations, such as `vc-relax` the base work
chain can implement handlers to deal with those instead of having them
trigger an unexpected failure. `PwCalculation` codes now handled:

 * ERROR_IONIC_CONVERGENCE_REACHED_EXCEPT_IN_FINAL_SCF
 * ERROR_IONIC_CONVERGENCE_NOT_REACHED
 * ERROR_IONIC_CYCLE_EXCEEDED_NSTEP
 * ERROR_IONIC_CYCLE_BFGS_HISTORY_FAILURE
 * ERROR_IONIC_CYCLE_BFGS_HISTORY_AND_FINAL_SCF_FAILURE
 * ERROR_IONIC_CYCLE_ELECTRONIC_CONVERGENCE_NOT_REACHED
 * ERROR_IONIC_CONVERGENCE_REACHED_FINAL_SCF_FAILED

Also the `ERROR_OUT_OF_WALLTIME` is updated to only restart from the
previous calculation if the structure has not changed. Because in the
latter case, often the restart will fail because the integrated charge
will no longer be correct.